refactor(backup/gcs): add `ManifestManager` for encapsulation
